Improve the PID drive command
